feat: DSTEW-1361 - Split out the application service into use case services, update controller calling and test
